CA

Description

Main functionality relating to cellular automata.

About comonadic cellular automata

What are Comonads?

A Comonad is a class which can be thought of as being the 'opposite' of a Monad. Whereas a Monad can be thought of as encapsulating the notion of an 'effectful computation' - that is, a computation which produces an effect - a Comonad can be thought of as encapsulating the notion of a 'contextual computation' - that is, a computation which takes place in a context, which can then be accessed during the computation. The methods of a Comonad are 'reversed' (or dual) compared to those of a Monad, as can be seen by comparing them:

return :: Monad m => a -> m a wraps a value in a monadic effect, whereas extract :: Comonad w => w a -> a isolates a value from its comonadic context

(>>=) :: Monad m => m a -> (a -> m b) -> m b applies an effectful computation a -> m b to a value m a, whereas flip extend :: Comonad w => w a -> (w a -> b) -> w b applies a contextual computation w a -> b to a value w a.

The stereotypical example of a comonad is the product comonad, which is just a partially applied tuple (,) r. This is the simplest comonad, representing a value with another value of type r acting as a context. If you said that this sounds like the Reader monad you would be right; this comonad is indeed isomorphic to the reader monad! In further detail:

extract :: (r, a) -> a removes the r context to leave just the a value; and

extend :: ((r, a) -> b) -> (r, a) -> (r, b) applies a contextual computation (r, a) -> b to a value (r, a).

Using Comonads to Simulate Cellular Automata

It turns out that comonads are a natural fit for cellular automata. The trick is to figure out how to view a cellular automaton as a contextual computation. If you have a look at a cellular automaton like Conway's Game of Life, the rules all follow the same general pattern of 'take a cell, and if the surrounding grid satisfies a certain condition then update the cell as follows'. This already is a contextful computation, in which the current cell is updated with reference to the context of the rest of the grid! This means that a cellular automaton comonad can be implemented as a pair of a 2D grid and one 'focused' point, to provide the 'current cell' referred to above (by the way, you might recognise this construction as a 2D list zipper). This is implemented as the Universe comonad below. A cellular automaton can then be implemented as a function Universe a -> b, the argument of which is pre-'focused' to the cell which is currently being recomputed. The extract method can be used to get the focused value, and the extend method can be used to apply a cellular automaton to a Universe. Putting this all together, the Game of Life can be implemented as follows:

data State = Alive | Dead
conwayLife :: Universe State -> State
conwayLife u = case extract u of                      -- We use extract to get the currently focused value
  Dead ->
    if count (experiment mkNbhd u) == 3               -- experiment is useful to get a neighbourhood from the focused cell
    then Alive
    else Dead
  Alive ->
    if count (experiment mkNbhd u) elem [2, 3]
    then Alive
    else Dead
 where
  mkNbhd (Point x y) = [ Point (x-1) (y-1)
                       , Point  x    (y-1)
                       , Point (x+1) (y-1)
                       , Point (x-1)  y
                       , Point (x+1)  y
                       , Point (x-1) (y+1)
                       , Point  x    (y+1)
                       , Point (x+1) (y+1)
                       ]
 count = sum . fmap (case { Alive -> 1 ; Dead -> 0 }) -- Note that this uses -XLambdaCase

Note that it would normally be much easier to use the functions from CA.Utils, but in the interests of pedagogy these have been avoided.

Convenient re-exports

Types

data Axis

A type used with -XDataKinds to parameterise Coord.

Constructors

X

Y

newtype Coord (x :: Axis)

Constructors

Coord

Fields

getCoord :: Int

data Point

Constructors

Point (Coord X) (Coord Y)

data Universe a

The type of a finite, 2D universe of cells of type a.

Internally, this is stored as a list of columns and a single 'focused' point, to form a 2D zipper.

type CARule g a = Universe a -> Rand g a

A convenient type signature for a stochastic cellular automaton update function.

data Bounds

Specifies a portion of a Universe.

Constructors

Bounds

Fields

boundsLeft :: !(Coord X)

boundsRight :: !(Coord X)

boundsTop :: !(Coord Y)

boundsBottom :: !(Coord Y)

boundsWidth :: Bounds -> Coord X

Finds the number of columns in a Bounds.

boundsHeight :: Bounds -> Coord Y

Finds the number of rows in a Bounds.

Functions on Universe

fromList :: [[a]] -> Universe a

Converts a list of columns to a Universe.

size :: Universe a -> (Coord X, Coord Y)

Returns the width and height of a Universe.

evolve :: CARule g a -> Universe a -> Rand g (Universe a)

Applies a stochastic evolution function to a Universe.

modifyPoint :: Point -> (a -> a) -> Universe a -> Universe a

Changes the value of a single Point in a Universe.

TODO The behaviour is undefined in the case when the point is outside the edges of the Universe. This is a bug and will be fixed later.

Other functions which Haddock can't document properly

extend :: (Universe a -> b) -> Universe a -> Universe b

This function can be used to evolve a Universe by one generation using a non-stochastic evolution function.

extract :: Universe a -> a

Returns the current focused point.

experiment :: Functor f => (Point -> f Point) -> Universe a -> f a

Given a function which turns the current focused point into a functor (e.g. a list) of points, experiment returns the values at these points. Useful with the neighbourhood functions from CA.Utils.

Conversion from and to lists

render :: Universe a -> [[a]]

Converts a Universe to a list of columns.

clip :: Universe a -> Bounds -> [[a]]

Extracts a portion of a Universe to a list of columns. Wraps around toroidally if a point outside the edges of the universe has been requested.

clipInside :: Universe a -> Bounds -> (Bounds, [[a]])

Similar to clip, but instead of 'wrapping around' toroidally if a point outside the edges of the universe has been requested, it stops at the edge of the universe. This means that the list returned may not cover all of the bounds requested, so clipInside also returns the bounds which were actually returned. For example, if the universe extends from the point (0,0) to (20, 20) but Bounds{boundsLeft=15, boundsRight=25, boundsTop=15, boundsBottom=25} was requested, then the portion outside the universe will be ignored and clipInside will behave as if only Bounds{boundsLeft=15, boundsRight=20, boundsTop=15, boundsBottom=20} was requested, and will return this value as the first part of the tuple.