CA.ALPACA

Description

Implements a parser for version 1.0 of the ALPACA language. Documentation is available at https://github.com/catseye/ALPACA/blob/fcd5a8bdd579c209475dedd2e221b8c437d5ca21/doc/ALPACA.markdown

Documentation

runALPACA :: RandomGen g => String -> Either String (CARule g Int, [Int])

Converts an ALPACA specification to a CARule. The CARule returned operates on an Int state; the allowed values are specified in the second element of the returned tuple.

TODO: Finite from the 'finite-typelits' package might be a better fit for the state than Int.