CA.Format.MCell

Description

Encoding and decoding the MCell file format.

Types

data MCell

A record containing all elements of the MCell format. For details on what each field means see http://psoup.math.wisc.edu/mcell/ca_files_formats.html#MCell

Constructors

MCell

Fields

game :: Maybe Game

rule :: Maybe String

speed :: Maybe Int

ccolors :: Maybe Int

coloring :: Maybe Coloring

wrap :: Maybe Bool

palette :: Maybe String

description :: Maybe String

universe :: Universe Int

diversities :: [(String, String)]

mcell :: Universe Int -> [(String, String)] -> MCell

A convenient constructor for an MCell value. Since there are only two non-Maybe fields in MCell, this constructor can be used to set all other fields to Nothing, after which other fields can be modified as desired using record update syntax. Usage is something like e.g. (mcell myUn myDiv){ game = Generations }.

data Game

Constructors

Life

Generations

WeightedLife

VoteForLife

RulesTable

CyclicCA

OneDTotalistic

NeumannBinary

GeneralBinary

LargerThanLife

Margolus

UserDLL

SpecialRules

data Coloring

Constructors

Standard

Alternate

Encoding and decoding

encodeMCell :: MCell -> String

decodeMCell :: String -> Either String MCell