CA.Internal

Description

A module containing the internals of CA. Do not use unless you really need to!

Convenient re-exports

Types

data Axis

A type used with -XDataKinds to parameterise Coord.

Constructors

X

Y

newtype Coord (x :: Axis)

Constructors

Coord

Fields

getCoord :: Int

data Point

Constructors

Point (Coord X) (Coord Y)

data Universe a

The type of a finite, 2D universe of cells of type a.

Internally, this is stored as a list of columns and a single 'focused' point, to form a 2D zipper.

Constructors

Universe (Seq (Seq a)) Point

type CARule g a = Universe a -> Rand g a

A convenient type signature for a stochastic cellular automaton update function.

data Bounds

Specifies a portion of a Universe.

Constructors

Bounds

Fields

boundsLeft :: !(Coord X)

boundsRight :: !(Coord X)

boundsTop :: !(Coord Y)

boundsBottom :: !(Coord Y)

boundsWidth :: Bounds -> Coord X

Finds the number of columns in a Bounds.

boundsHeight :: Bounds -> Coord Y

Finds the number of rows in a Bounds.

Functions on Universe

size :: Universe a -> (Coord X, Coord Y)

Returns the width and height of a Universe.

normalize :: Universe a -> Universe a

Wraps the coordinates of the focus to within the allowed range. Generally you won't need to use this unless you're messing around with the internals of a Universe.

evolve :: CARule g a -> Universe a -> Rand g (Universe a)

Applies a stochastic evolution function to a Universe.

modifyPoint :: Point -> (a -> a) -> Universe a -> Universe a

Changes the value of a single Point in a Universe.

TODO The behaviour is undefined in the case when the point is outside the edges of the Universe. This is a bug and will be fixed later.

Conversion from and to lists

fromList :: [[a]] -> Universe a

Converts a list of columns to a Universe.

render :: Universe a -> [[a]]

Converts a Universe to a list of columns.

clip :: Universe a -> Bounds -> [[a]]

Extracts a portion of a Universe to a list of columns. Wraps around toroidally if a point outside the edges of the universe has been requested.

clipInside :: Universe a -> Bounds -> (Bounds, [[a]])

Similar to clip, but instead of 'wrapping around' toroidally if a point outside the edges of the universe has been requested, it stops at the edge of the universe. This means that the list returned may not cover all of the bounds requested, so clipInside also returns the bounds which were actually returned. For example, if the universe extends from the point (0,0) to (20, 20) but Bounds{boundsLeft=15, boundsRight=25, boundsTop=15, boundsBottom=25} was requested, then the portion outside the universe will be ignored and clipInside will behave as if only Bounds{boundsLeft=15, boundsRight=20, boundsTop=15, boundsBottom=20} was requested, and will return this value as the first part of the tuple.