CA.Utils

Description

Utility functions relating to cellular automata.

Neighbourhoods

For definitions of the neighbourhoods below, see https://en.wikipedia.org/wiki/Moore_neighborhood and https://en.wikipedia.org/wiki/Von_Neumann_neighborhood.

moore

Arguments

:: Bool
Whether to include the central point

-> Point -> [Point]

Generates a Moore neighbourhood for use with experiment.

vonNeumann

Arguments

:: Bool
Whether to include the central point

-> Point -> [Point]

Generates a von Neumann neighbourhood for use with experiment.

pointed :: [Point] -> Point -> [Point]

Turns a list of points into a neighbourhood which can be used with experiment.

count :: (Functor t, Foldable t) => (a -> Bool) -> t a -> Int

Counts the number of values in a Foldable satisfying a condition. Useful with neighbourhoods e.g. counting the number of live cells in a Moore neighbourhood using count (==Alive) . experiment (moore False)

Other functions

data Direction

A direction - either left, right, up or down. Each direction is suffixed with Dir to avoid confusion with Left or Right.

Constructors

LeftDir

RightDir

DownDir

UpDir

move :: Direction -> Point -> Point

Translates a Point by a Direction.