Hint.Interop

Description

Types necessary for communication between Hint and the application.

Documentation

data CAVals' t

A record containing all the various bits and pieces of information which the application needs to know to load a new rule.

Constructors

CAVals'

Fields

_rule :: CARule StdGen t

The rule itself

_states :: [t]

The states which can be selected from the state selection menu

_defaultPattern :: Universe t

The default pattern displayed before anything has been edited

_state2color :: t -> (Double, Double, Double)

A function to convert states into (red, green, blue) colours which are displayed on the grid

_encodeInt :: t -> Int

Encodes a state into an integer which must be between 0 and 255. Used to save a pattern to a file.

_decodeInt :: Int -> t

Decodes a state from an integer between 0 and 255. Used to load a pattern from a file.

data CAVals

If we decide to use a custom state type for our CA, we can't use that type outside of Hint, so this existential wrapper is used to 'hide' the state type from the outside world.

Constructors

Eq t => CAVals (CAVals' t)